Move each stats exporter into its own directory. (#48)
