+++ opencensus/exporters/stats/stackdriver/internal/stackdriver_utils.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "opencensus/exporters/stats/internal/stackdriver_utils.h"
+#include "opencensus/exporters/stats/stackdriver/internal/stackdriver_utils.h"
 
 #include <string>
